Fixed the cube
The problem with the cube's lighting model is that a cube shouldn't act
like that for a directional light. If you light the cube from the top,
only the top face should be lit as there is no bounce effect to reflect
light back onto the sides of the cube, the other faces are either
perpendicular or turned away from the light source so there is no chance
of light hitting those faces

Therefore, instead of composing the cube out of 8 vertices, I've instead
composed it out of 24 vertices, four for each face. This is also
essential as we can then set the colour of each face separately to the
rest. The normals for each face point directly out from the face.
Lighting now works properly