wrote up a function that find the velocity value.
Finds the velocity level depending on the distance of the tap position
and the hero position.

- There is a max_distance which is basically a restriction to how far
the player need to tap within the hero to have an movement.
- There is a speed value which is the default value of speed that is
used for movement.
For the speed and max_distance, it needs testing we can change these
value and set it so that the movement is smooth.

Problem: Im not sure how to do the part where we actually decrease the
velocity per frame so that the character is actually moving to a stop
after each tap.